+++ src/Symfony/Component/Security/Http/CHANGELOG.md
@@ -7,6 +7,7 @@ CHANGELOG
  * Pass the current token to the `checkPostAuth()` method of user checkers
  * Deprecate argument `$secret` of `RememberMeAuthenticator`
  * Deprecate passing an empty string as `$userIdentifier` argument to `UserBadge` constructor
+ * Add encryption support to `OidcTokenHandler` (JWE)
 
 7.1
 ---